Fix(stations): Geocode stations and improve fixing scripts
This commit addresses the issue of incorrect station locations, including a large grid of stations in the Atlantic Ocean.

The key changes are:
- Consolidated multiple redundant station-fixing scripts into a single, robust script: `fixAndGeocodeStations.js`.
- Updated `package.json` to use this new script for `update-station-coordinates` and `verify-station-coordinates`.
- The script now correctly targets the `...Fixed.json` files that the application uses.
- The script uses the OpenStreetMap Nominatim API to geocode stations with invalid coordinates.
- The script has been improved to detect more classes of invalid coordinates, including the Atlantic grid pattern.
- A function has been added to extract location information from station names to improve geocoding accuracy.
- Deleted old, confusing, and redundant scripts.
- The station data files have been updated with the corrected coordinates.
