+++ fixAndGeocodeStations.js
@@ -1,239 +1,347 @@
-import fs from 'fs';
-import path from 'path';
+import fs from 'fs/promises';
 import axios from 'axios';
 
-const TV_STATIONS_PATH = path.join('src', 'data', 'tvStationsWithUrls.json');
-const RADIO_STATIONS_PATH = path.join('src', 'data', 'radioStations.json');
-const BACKUP_DIR = 'backups';
-
-// --- Manual Override Data ---
-
-const corrected_tv_locations = {
-    "KERODT1.us": { "city": "Bakersfield", "country": "USA" },
-    "360RFTV.cr": { "city": "Nosara", "country": "Costa Rica" },
-    "APlusGuate.gt": { "city": "Guatemala City", "country": "Guatemala" },
-    "ABCAustralia.au@Vietnam": { "city": "Sydney", "country": "Australia" },
-    "AbuDhabiAloula.ae@SD": { "city": "Abu Dhabi", "country": "UAE" },
-    "AbuDhabiEmirates.ae@SD": { "city": "Abu Dhabi", "country": "UAE" },
-    "ACTV.be": { "city": "Wallonia", "country": "Belgium" },
-    "ACWUGTV.ug": { "city": "Kampala", "country": "Uganda" },
-    "AlAoula.ma@Inter": { "city": "Rabat", "country": "Morocco" },
-    "LaayouneTV.ma@SD": { "city": "Laayoune", "country": "Morocco" },
-    "AlJadeed.lb@SD": { "city": "Beirut", "country": "Lebanon" },
-    "AlMaghribia.ma@SD": { "city": "Rabat", "country": "Morocco" },
-    "AlShoub.eg@SD": { "city": "Cairo", "country": "Egypt" },
-    "AltantoTV.do": { "city": "Santo Domingo", "country": "Dominican Republic" },
-    "AltynAsyr.tm": { "city": "Ashgabat", "country": "Turkmenistan" },
-    "AmericaTeVe.pr": { "city": "San Juan", "country": "Puerto Rico" },
-    "AmouzeshTV.ir@SD": { "city": "Tehran", "country": "Iran" },
-    "Antena3Internacional.es@SD": { "city": "Madrid", "country": "Spain" },
-    "Asgabat.tm": { "city": "Ashgabat", "country": "Turkmenistan" },
-    "AstroVaanavil.my@SD": { "city": "Kuala Lumpur", "country": "Malaysia" },
-    "ATBLaPaz.bo": { "city": "La Paz", "country": "Bolivia" },
-    "aurLifeHD.pk": { "city": "Karachi", "country": "Pakistan" },
-    "AztecaGuatemala.gt": { "city": "Guatemala City", "country": "Guatemala" },
-    "AztecaHonduras.hn": { "city": "Tegucigalpa", "country": "Honduras" },
-    "BahrainInternational.bh": { "city": "Manama", "country": "Bahrain" },
-    "BBCFourCBeebies.uk@HD": { "city": "London", "country": "UK" },
-    "BBCOne.uk@ChannelIslands": { "city": "Saint Helier", "country": "Jersey" },
-    "BBCOne.uk@YorkshireHD": { "city": "Leeds", "country": "UK" },
-    "BBCOne.uk@EastMidlands": { "city": "Nottingham", "country": "UK" },
-    "BBCOne.uk@EastMidlandsHD": { "city": "Nottingham", "country": "UK" },
-    "BBCOne.uk@Yorkshire": { "city": "Hull", "country": "UK" },
-    "BBCOne.uk@NorthEastCumbria": { "city": "Newcastle upon Tyne", "country": "UK" },
-    "BBCOne.uk@NorthEastCumbriaHD": { "city": "Newcastle upon Tyne", "country": "UK" },
-    "BBCOne.uk@NorthWest": { "city": "Manchester", "country": "UK" },
-    "BBCOne.uk@NorthWestHD": { "city": "Manchester", "country": "UK" },
-    "BBCOne.uk@NorthernIreland": { "city": "Belfast", "country": "UK" },
-    "BBCOne.uk@NorthernIrelandHD": { "city": "Belfast", "country": "UK" },
-    "BBCOne.uk@Scotland": { "city": "Glasgow", "country": "UK" },
-    "BBCOne.uk@ScotlandHD": { "city": "Glasgow", "country": "UK" },
-    "BBCOne.uk@South": { "city": "Southampton", "country": "UK" },
-    "BBCOne.uk@SouthEast": { "city": "Tunbridge Wells", "country": "UK" },
-    "BBCOne.uk@SouthWest": { "city": "Plymouth", "country": "UK" },
-    "BBCOne.uk@SouthWestHD": { "city": "Plymouth", "country": "UK" },
-    "BBCOne.uk@Wales": { "city": "Cardiff", "country": "UK" },
-    "BBCOne.uk@West": { "city": "Bristol", "country": "UK" },
-    "BBCOne.uk@WestMidlands": { "city": "Birmingham", "country": "UK" },
-    "BBCOne.uk@YorkshireLincolnshire": { "city": "Leeds", "country": "UK" },
-    "BBCThree.uk@HD": { "city": "London", "country": "UK" },
-    "BBCTwo.uk@England": { "city": "London", "country": "UK" },
-    "BBCTwo.uk@HD": { "city": "London", "country": "UK" },
-    "BBCTwo.uk@NorthernIrelandHD": { "city": "Belfast", "country": "UK" },
-    "BBCTwo.uk@Wales": { "city": "Cardiff", "country": "UK" },
-    "BCSStarCrossTV.ng@SD": { "city": "Lagos", "country": "Nigeria" },
-    "Bolivision.bo": { "city": "La Paz", "country": "Bolivia" },
-    "Canal4RD.do": { "city": "Santo Domingo", "country": "Dominican Republic" },
-    "Canal13Esquipulas.gt": { "city": "Esquipulas", "country": "Guatemala" },
-    "Canal25.do@SD": { "city": "Santo Domingo", "country": "Dominican Republic" },
-    "CanalMacau.mo": { "city": "Macau", "country": "China" },
-    "CatveFM.br": { "city": "Cascavel", "country": "Brazil" },
-    "CBC.eg": { "city": "Cairo", "country": "Egypt" }
-};
-
-const corrected_radio_locations = {
-    "talkSPORT": { "city": "London", "country": "UK" },
-    "talkRADIO": { "city": "London", "country": "UK" },
-    "Радио Маяк (Radio Mayak)": { "city": "Moscow", "country": "Russia" },
-    "Дорожное радио": { "city": "Moscow", "country": "Russia" },
-    "Комсомольская правда": { "city": "Moscow", "country": "Russia" },
-    "Radio Record - Russian Mix": { "city": "Saint Petersburg", "country": "Russia" },
-    "Радио Русские Песни": { "city": "Moscow", "country": "Russia" },
-    "BBC Radio 6 Music": { "city": "London", "country": "UK" },
-    "Radio Mars": { "city": "Casablanca", "country": "Morocco" },
-    "CNN": { "city": "Atlanta", "country": "USA" },
-    "MSNBC": { "city": "New York", "country": "USA" },
-    "NPR 24 Hour Program Stream": { "city": "Washington D.C.", "country": "USA" }
-};
-
-// --- Helper Functions ---
-
-const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
-
-const backupFile = (filePath) => {
-  if (!fs.existsSync(BACKUP_DIR)) {
-    fs.mkdirSync(BACKUP_DIR);
+// Function to check if coordinates are valid
+function isValidCoordinate(lat, lon) {
+  // Check if coordinates are numbers
+  if (typeof lat !== 'number' || typeof lon !== 'number') {
+    return false;
   }
-  const timestamp = new Date().toISOString().replace(/[:.]/g, '-');
-  const backupPath = path.join(BACKUP_DIR, `${timestamp}-${path.basename(filePath)}`);
-  fs.copyFileSync(filePath, backupPath);
-  console.log(`Backup of ${filePath} created at ${backupPath}`);
-};
 
-// --- Geocoding ---
+  // Check if coordinates are within valid ranges
+  const validLat = lat >= -90 && lat <= 90;
+  const validLon = lon >= -180 && lon <= 180;
 
-const GEOCODE_CACHE = new Map();
+  // Check if coordinates are not placeholder values
+  const isPlaceholder =
+    (lat === 0 && lon === 0) ||
+    (lat === 40 && lon === -100) ||
+    (lat === 42 && lon === -100) ||
+    (lat === 12.4989994 && lon === 124.6746741) ||
+    (lat === 40 && lon === -98) ||
+    (lat === 42 && lon === -98) ||
+    (lat === 44 && lon === -98) ||
+    (lat === 46 && lon === -98) ||
+    (lat === 48 && lon === -98) ||
+    (lat === 50 && lon === -98) ||
+    (lat === 52 && lon === -98) ||
+    (lat === 54 && lon === -98) ||
+    (lat === 56 && lon === -98) ||
+    (lat === 58 && lon === -98) ||
+    // Additional placeholder patterns
+    (Math.abs(lat) < 1 && Math.abs(lon) < 1) || // Very close to (0,0)
+    (lat > 20 && lat < 70 && lon > -100 && lon < -20); // Atlantic grid
 
-async function geocodeLocation(query) {
-  if (GEOCODE_CACHE.has(query)) {
-    return GEOCODE_CACHE.get(query);
-  }
-  try {
-    console.log(`Geocoding: "${query}"`);
-    const response = await axios.get('https://nominatim.openstreetmap.org/search', {
-      params: { q: query, format: 'json', limit: 1, 'accept-language': 'en' },
-      headers: { 'User-Agent': 'Cool-Station-Fixer/1.0' }
-    });
-    await sleep(1100);
-    if (response.data && response.data.length > 0) {
-      const { lat, lon, display_name } = response.data[0];
-      const address = response.data[0].address;
-      const country = address?.country || '';
-      const city = address?.city || address?.town || address?.village || address?.hamlet || '';
-      const result = { lat: parseFloat(lat), lon: parseFloat(lon), displayName: display_name, country: country, city: city };
-      GEOCODE_CACHE.set(query, result);
-      return result;
+  return validLat && validLon && !isPlaceholder;
+}
+
+// Function to extract location from station name
+function extractLocationFromName(name) {
+    // Regex for City, ST where ST is a valid US state
+    const states = ["AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DE", "FL", "GA",
+                    "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD",
+                    "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ",
+                    "NM", "NY", "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC",
+                    "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY"];
+    const cityStateRegex = new RegExp(`([A-Za-z\\s]+?),?\\s(${states.join('|')})\\b`);
+    const match = name.match(cityStateRegex);
+
+    if (match && match[1] && match[2]) {
+        return { city: match[1].trim(), state: match[2].trim() };
     }
+
+    // Add more regex patterns here for other formats if needed
+
+    return null;
+}
+
+// Function to geocode a location using OpenStreetMap Nominatim
+async function geocodeLocation(city, state, country) {
+  try {
+    // Handle special cases or unknown locations
+    if ((!city || city === 'Unknown') && (!state || state === 'Unknown') && (!country || country === 'Unknown')) {
+      return null;
+    }
+
+    // Build location query with available information
+    const locationParts = [];
+    if (country && country !== 'Unknown' && country.trim() !== '') {
+      locationParts.push(country.trim());
+    }
+    if (state && state !== 'Unknown' && state.trim() !== '') {
+      locationParts.push(state.trim());
+    }
+    if (city && city !== 'Unknown' && city.trim() !== '') {
+      locationParts.push(city.trim());
+    }
+
+    if (locationParts.length === 0) {
+      return null;
+    }
+
+    const locationQuery = locationParts.join(', ');
+
+    console.log(`Geocoding location: ${locationQuery}`);
+
+    // Use OpenStreetMap Nominatim API for geocoding
+    const response = await axios.get('https://nominatim.openstreetmap.org/search', {
+      params: {
+        q: locationQuery,
+        format: 'json',
+        limit: 1
+      },
+      headers: {
+        'User-Agent': 'GlobeMediaStreamer/1.0 (contact@example.com)'
+      },
+      timeout: 10000
+    });
+
+    if (response.data && response.data.length > 0) {
+      const result = response.data[0];
+      return {
+        latitude: parseFloat(result.lat),
+        longitude: parseFloat(result.lon)
+      };
+    }
+
     return null;
   } catch (error) {
-    console.error(`Error geocoding "${query}":`, error.message);
-    await sleep(1100);
+    console.warn(`Geocoding failed for ${city}, ${state}, ${country}:`, error.message);
     return null;
   }
 }
 
-// --- Location Logic ---
+// Function to find stations with invalid coordinates
+async function findInvalidStations(data, dataType) {
+  console.log(`\nChecking ${data.length} ${dataType} stations for invalid coordinates...`);
 
-function isInvalidLocation(station) {
-    const lat = parseFloat(station.geo_lat);
-    const lon = parseFloat(station.geo_long);
-    const inAtlanticBox = lat > 20 && lat < 70 && lon > -100 && lon < -10; // Bounding box for incorrectly placed stations
-    if (isNaN(lat) || isNaN(lon) || (lat === 0 && lon === 0) || inAtlanticBox) {
-        return true;
+  const invalidStations = [];
+
+  for (let i = 0; i < data.length; i++) {
+    const station = data[i];
+    const hasValidCoords = isValidCoordinate(station.geo_lat, station.geo_long);
+
+    if (!hasValidCoords) {
+      invalidStations.push({
+        index: i,
+        station: station
+      });
     }
-    if (station.country === 'Unknown') {
-        return true;
-    }
-    return false;
+  }
+
+  console.log(`Found ${invalidStations.length} ${dataType} stations with invalid coordinates.`);
+  return invalidStations;
 }
 
-function getGeocodeQuery(station, isRadio = false) {
-    const override_map = isRadio ? corrected_radio_locations : corrected_tv_locations;
-    const key = isRadio ? station.name : station.id;
+// Function to fix station coordinates
+async function fixStationCoordinates(station, geocodedCache) {
+  // Create a cache key
+  const cacheKey = `${station.city || ''}|${station.state || ''}|${station.country || ''}`;
 
-    if (override_map[key]) {
-        const { city, country } = override_map[key];
-        return `${city}, ${country}`;
+  // Check if we already have this location geocoded
+  if (geocodedCache.has(cacheKey)) {
+    const coords = geocodedCache.get(cacheKey);
+    if (coords) {
+      return {
+        success: true,
+        coordinates: coords,
+        usedCache: true
+      };
+    } else {
+      return {
+        success: false,
+        usedCache: true
+      };
     }
+  }
 
-    if (station.city !== 'Unknown' && station.country !== 'Unknown') {
-        return `${station.city}, ${station.country}`;
-    }
+  // Try to extract location from name if city is not reliable
+  let extractedLocation = null;
+  if (!station.city || station.city === 'Unknown' || station.city.includes(station.name.substring(0, 10))) {
+      extractedLocation = extractLocationFromName(station.name);
+  }
 
-    let query = station.city !== 'Unknown' ? station.city : station.name;
-    if (typeof query !== 'string') return null;
+  let city = station.city;
+  let state = station.state;
 
-    const cityStateMatch = query.match(/\b([A-Za-z\s]+?)\s([A-Z]{2})\b/);
-    if (cityStateMatch && cityStateMatch[2] !== 'TV') return `${cityStateMatch[1]}, ${cityStateMatch[2]}, USA`;
+  if(extractedLocation) {
+    console.log(`Extracted location for ${station.name}: ${extractedLocation.city}, ${extractedLocation.state}`);
+    city = extractedLocation.city;
+    state = extractedLocation.state;
+  }
 
-    if (query.includes('BBC')) {
-        const regionMatch = query.match(/BBC\s(?:One|Two|Four|Scotland|Alba|Three)\s(.+?)(?:\sHD|\s\(|\s$)/);
-        if (regionMatch && regionMatch[1]) return `${regionMatch[1]}, UK`;
-    }
+  // Try to geocode this location
+  console.log(`Geocoding ${station.name} (${city}, ${state}, ${station.country})...`);
+  const coords = await geocodeLocation(city, state, station.country);
+  geocodedCache.set(cacheKey, coords);
 
-    const tldMatch = station.id && station.id.match(/\.([a-z]{2})$/);
-    if (tldMatch) {
-        const countryMap = { de: 'Germany', fr: 'France', it: 'Italy', es: 'Spain', uk: 'United Kingdom', nl: 'Netherlands', za: 'South Africa', af: 'Afghanistan', ge: 'Georgia', ru: 'Russia', th: 'Thailand', ar: 'Argentina', sn: 'Senegal', mx: 'Mexico', hu: 'Hungary', us: 'USA', ci: 'Ivory Coast', tr: 'Turkey', ro: 'Romania', ma: 'Morocco', ps: 'Palestine', kr: 'South Korea', ae: 'United Arab Emirates', ug: 'Uganda', cy: 'Cyprus', et: 'Ethiopia', br: 'Brazil', gh: 'Ghana', iq: 'Iraq', ca: 'Canada', lb: 'Lebanon', qa: 'Qatar', sa: 'Saudi Arabia', sy: 'Syria', az: 'Azerbaijan', ir: 'Iran', am: 'Armenia', gr: 'Greece', id: 'Indonesia', ua: 'Ukraine', cl: 'Chile', ve: 'Venezuela', np: 'Nepal', ec: 'Ecuador', py: 'Paraguay', cm: 'Cameroon', vn: 'Vietnam', dk: 'Denmark', ng: 'Nigeria', mn: 'Mongolia', co: 'Colombia', sv: 'El Salvador', ad: 'Andorra', at: 'Austria', gn: 'Guinea', pe: 'Peru', kz: 'Kazakhstan', au: 'Australia', pk: 'Pakistan', tj: 'Tajikistan', ba: 'Bosnia and Herzegovina', bd: 'Bangladesh', my: 'Malaysia', rw: 'Rwanda', jo: 'Jordan', lc: 'Saint Lucia' };
-        if (countryMap[tldMatch[1]]) return countryMap[tldMatch[1]];
-    }
-
-    if (station.city && station.city !== 'Unknown') return station.city;
-    if (station.country && station.country !== 'Unknown') return station.country;
-
-    return null;
-}
-
-// --- Main Processing ---
-
-async function processStations(filePath, isRadio = false) {
-  console.log(`\n--- Processing ${filePath} ---`);
-  const stations = JSON.parse(fs.readFileSync(filePath, 'utf-8'));
-  let fixedCount = 0;
-  let failedCount = 0;
-
-  for (let i = 0; i < stations.length; i++) {
-    const station = stations[i];
-    if (isInvalidLocation(station) || (isRadio ? corrected_radio_locations[station.name] : corrected_tv_locations[station.id])) {
-      const query = getGeocodeQuery(station, isRadio);
-      if (query) {
-        const geodata = await geocodeLocation(query);
-        if (geodata) {
-          console.log(`  [FIXED] ${station.name}: ${geodata.displayName}`);
-          station.geo_lat = geodata.lat;
-          station.geo_long = geodata.lon;
-          if (geodata.city) station.city = geodata.city;
-          if (geodata.country) station.country = geodata.country;
-          fixedCount++;
-        } else {
-          console.log(`  [FAILED] Could not geocode "${query}" for station: ${station.name}`);
-          failedCount++;
-        }
-      } else {
-          console.log(`  [SKIPPED] Could not parse location for station: ${station.name}`);
-          failedCount++;
+  if (coords) {
+    return {
+      success: true,
+      coordinates: coords,
+      usedCache: false
+    };
+  } else {
+    // Try with just the country if we have one
+    if (station.country && station.country !== 'Unknown') {
+      console.log(`Trying with just country: ${station.country}`);
+      const countryCoords = await geocodeLocation(null, null, station.country);
+      if (countryCoords) {
+        geocodedCache.set(cacheKey, countryCoords);
+        return {
+          success: true,
+          coordinates: countryCoords,
+          usedCache: false
+        };
       }
     }
-  }
 
-  fs.writeFileSync(filePath, JSON.stringify(stations, null, 2));
-  console.log(`--- Finished processing ${filePath} ---`);
-  console.log(`Fixed: ${fixedCount}, Failed/Skipped: ${failedCount}`);
-  return { fixedCount, failedCount };
+    return {
+      success: false,
+      usedCache: false
+    };
+  }
 }
 
-// --- Execution ---
+// Function to process and fix station data
+async function processStationData(filePath, dataType) {
+  try {
+    console.log(`\n=== Processing ${dataType} stations from ${filePath} ===`);
 
-(async () => {
-  console.log("Starting consolidated station coordinate correction process...");
+    // Read the current stations file
+    const stationsData = await fs.readFile(filePath, 'utf8');
+    let stations = JSON.parse(stationsData);
 
-  backupFile(TV_STATIONS_PATH);
-  backupFile(RADIO_STATIONS_PATH);
+    console.log(`Loaded ${stations.length} ${dataType} stations.`);
 
-  const tvResults = await processStations(TV_STATIONS_PATH, false);
-  const radioResults = await processStations(RADIO_STATIONS_PATH, true);
+    // Find invalid stations
+    const invalidStations = await findInvalidStations(stations, dataType);
 
-  console.log("\n--- Summary ---");
-  console.log(`TV Stations: ${tvResults.fixedCount} fixed, ${tvResults.failedCount} failed/skipped.`);
-  console.log(`Radio Stations: ${radioResults.fixedCount} fixed, ${radioResults.failedCount} failed/skipped.`);
-  console.log("Process complete.");
-})();
+    if (invalidStations.length === 0) {
+      console.log(`All ${dataType} stations have valid coordinates. No fixes needed.`);
+      return stations;
+    }
+
+    // Keep track of geocoded locations to avoid repeated API calls
+    const geocodedCache = new Map();
+
+    // Counter for stations that need geocoding
+    let geocodedStations = 0;
+    let skippedStations = 0;
+
+    // Process each invalid station
+    for (let i = 0; i < invalidStations.length; i++) {
+      const { index, station } = invalidStations[i];
+
+      console.log(`\nProcessing station ${i + 1}/${invalidStations.length}: ${station.name}`);
+
+      const result = await fixStationCoordinates(station, geocodedCache);
+
+      if (result.success) {
+        stations[index].geo_lat = result.coordinates.latitude;
+        stations[index].geo_long = result.coordinates.longitude;
+        geocodedStations++;
+
+        if (result.usedCache) {
+          console.log(`Using cached coordinates for ${station.name}: ${result.coordinates.latitude}, ${result.coordinates.longitude}`);
+        } else {
+          console.log(`Geocoded ${station.name}: ${result.coordinates.latitude}, ${result.coordinates.longitude}`);
+
+          // Add a small delay to respect API rate limits
+          await new Promise(resolve => setTimeout(resolve, 1000));
+        }
+      } else {
+        skippedStations++;
+        console.log(`Failed to geocode ${station.name}`);
+      }
+    }
+
+    console.log(`\n${dataType.toUpperCase()} STATIONS FIX SUMMARY:`);
+    console.log(`Successfully geocoded: ${geocodedStations}`);
+    console.log(`Skipped stations: ${skippedStations}`);
+
+    // Write the fixed data back to the file
+    await fs.writeFile(filePath, JSON.stringify(stations, null, 2));
+    console.log(`Fixed ${dataType} stations data saved to ${filePath}`);
+
+    return stations;
+  } catch (error) {
+    console.error(`Error processing ${dataType} station coordinates:`, error);
+    throw error;
+  }
+}
+
+// Main function to fix all station coordinates
+async function fixAllStationCoordinates() {
+  try {
+    console.log('=== STARTING COMPREHENSIVE STATION COORDINATE FIX ===');
+
+    // Process TV stations
+    await processStationData('./src/data/tvStationsWithUrlsFixed.json', 'TV');
+
+    // Process radio stations
+    await processStationData('./src/data/radioStationsFixed.json', 'radio');
+
+    console.log('\n=== ALL STATION COORDINATES FIXED SUCCESSFULLY ===');
+  } catch (error) {
+    console.error('Error fixing station coordinates:', error);
+  }
+}
+
+// Function to verify all station coordinates
+async function verifyAllStationCoordinates() {
+  try {
+    console.log('=== VERIFYING ALL STATION COORDINATES ===');
+
+    // Read and verify TV stations
+    const tvStationsData = await fs.readFile('./src/data/tvStationsWithUrlsFixed.json', 'utf8');
+    const tvStations = JSON.parse(tvStationsData);
+    const invalidTVStations = await findInvalidStations(tvStations, 'TV');
+
+    // Read and verify radio stations
+    const radioStationsData = await fs.readFile('./src/data/radioStationsFixed.json', 'utf8');
+    const radioStations = JSON.parse(radioStationsData);
+    const invalidRadioStations = await findInvalidStations(radioStations, 'radio');
+
+    console.log('\n=== VERIFICATION RESULTS ===');
+    console.log(`Invalid TV stations: ${invalidTVStations.length}`);
+    console.log(`Invalid radio stations: ${invalidRadioStations.length}`);
+
+    if (invalidTVStations.length === 0 && invalidRadioStations.length === 0) {
+      console.log('✅ All station coordinates are valid!');
+    } else {
+      console.log('❌ Some stations still have invalid coordinates.');
+
+      if (invalidTVStations.length > 0) {
+        console.log('\nInvalid TV stations:');
+        invalidTVStations.slice(0, 5).forEach(({ index, station }) => {
+          console.log(`  - Index ${index}: ${station.name} (${station.geo_lat}, ${station.geo_long})`);
+        });
+        if (invalidTVStations.length > 5) {
+          console.log(`  ... and ${invalidTVStations.length - 5} more`);
+        }
+      }
+
+      if (invalidRadioStations.length > 0) {
+        console.log('\nInvalid radio stations:');
+        invalidRadioStations.slice(0, 5).forEach(({ index, station }) => {
+          console.log(`  - Index ${index}: ${station.name} (${station.geo_lat}, ${station.geo_long})`);
+        });
+        if (invalidRadioStations.length > 5) {
+          console.log(`  ... and ${invalidRadioStations.length - 5} more`);
+        }
+      }
+    }
+  } catch (error) {
+    console.error('Error verifying station coordinates:', error);
+  }
+}
+
+// If this script is run directly, execute the fix function
+if (process.argv[1] && process.argv[1].endsWith('fixAndGeocodeStations.js')) {
+  if (process.argv.includes('--verify')) {
+    verifyAllStationCoordinates();
+  } else {
+    fixAllStationCoordinates();
+  }
+}
