+++ src/data/radioStationsFixed.json
@@ -1200,8 +1200,8 @@
     "tags": "country",
     "codec": "MP3",
     "bitrate": 128,
-    "geo_lat": 38.8888884,
-    "geo_long": -77.0405701,
+    "geo_lat": 36.1337324,
+    "geo_long": -80.2782636,
     "votes": 47925,
     "clickcount": 2593,
     "lastcheckok": 1,
@@ -1948,8 +1948,8 @@
     "tags": "classical,relax",
     "codec": "MP3",
     "bitrate": 128,
-    "geo_lat": 29.6264964,
-    "geo_long": -82.3645914,
+    "geo_lat": 49.3829708,
+    "geo_long": -95.1533219,
     "votes": 47899,
     "clickcount": 1931,
     "lastcheckok": 1,
@@ -2344,8 +2344,8 @@
     "tags": "80s,90s,hits,today",
     "codec": "MP3",
     "bitrate": 128,
-    "geo_lat": 38.8888884,
-    "geo_long": -77.0405701,
+    "geo_lat": 36.1337324,
+    "geo_long": -80.2782636,
     "votes": 37716,
     "clickcount": 1819,
     "lastcheckok": 1,
@@ -2432,8 +2432,8 @@
     "tags": "sport",
     "codec": "MP3",
     "bitrate": 64,
-    "geo_lat": 45,
-    "geo_long": -50,
+    "geo_lat": 52.4577754,
+    "geo_long": -1.8692099,
     "votes": 1072,
     "clickcount": 1792,
     "lastcheckok": 1,
@@ -2674,8 +2674,8 @@
     "tags": "talk",
     "codec": "MP3",
     "bitrate": 64,
-    "geo_lat": 25,
-    "geo_long": -40,
+    "geo_lat": 52.4577754,
+    "geo_long": -1.8692099,
     "votes": 15151,
     "clickcount": 1752,
     "lastcheckok": 1,
@@ -2762,8 +2762,8 @@
     "tags": "music,news,talk",
     "codec": "MP3",
     "bitrate": 192,
-    "geo_lat": 45,
-    "geo_long": -40,
+    "geo_lat": 64.6863136,
+    "geo_long": 97.7453061,
     "votes": 38009,
     "clickcount": 1721,
     "lastcheckok": 1,
@@ -2784,8 +2784,8 @@
     "tags": "dorojnoe",
     "codec": "MP3",
     "bitrate": 320,
-    "geo_lat": 50,
-    "geo_long": -40,
+    "geo_lat": 59.9357842,
+    "geo_long": 30.3002164,
     "votes": 47939,
     "clickcount": 1720,
     "lastcheckok": 1,
@@ -2828,8 +2828,8 @@
     "tags": "news,talk",
     "codec": "MP3",
     "bitrate": 0,
-    "geo_lat": 60,
-    "geo_long": -40,
+    "geo_lat": 55.7497417,
+    "geo_long": 37.6022613,
     "votes": 7656,
     "clickcount": 1713,
     "lastcheckok": 1,
@@ -3092,8 +3092,8 @@
     "tags": "",
     "codec": "AAC",
     "bitrate": 0,
-    "geo_lat": 45,
-    "geo_long": -30,
+    "geo_lat": 64.6863136,
+    "geo_long": 97.7453061,
     "votes": 40272,
     "clickcount": 1625,
     "lastcheckok": 1,
@@ -3114,8 +3114,8 @@
     "tags": "russian pop",
     "codec": "MP3",
     "bitrate": 128,
-    "geo_lat": 50,
-    "geo_long": -30,
+    "geo_lat": 64.6863136,
+    "geo_long": 97.7453061,
     "votes": 47917,
     "clickcount": 1622,
     "lastcheckok": 1,
