+++ src/data/tvStationsWithUrlsFixed.json
@@ -92,8 +92,8 @@
     "city": "2M",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 25.029422,
-    "geo_long": -77.361956,
+    "geo_lat": 58.2287982,
+    "geo_long": 6.5242929,
     "latitude": 25.029422,
     "longitude": -77.361956
   },
@@ -246,8 +246,8 @@
     "city": "8N",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 25.029422,
-    "geo_long": -77.361956,
+    "geo_lat": 43.2430316,
+    "geo_long": 5.378865,
     "latitude": 25.029422,
     "longitude": -77.361956
   },
@@ -512,8 +512,8 @@
     "city": "A2",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 25.029422,
-    "geo_long": -77.361956,
+    "geo_lat": 47.1340383,
+    "geo_long": 8.2056284,
     "latitude": 25.029422,
     "longitude": -77.361956
   },
@@ -540,8 +540,8 @@
     "city": "A12",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 25.029422,
-    "geo_long": -77.361956,
+    "geo_lat": 42.0066871,
+    "geo_long": 12.0308307,
     "latitude": 25.029422,
     "longitude": -77.361956
   },
@@ -596,8 +596,8 @@
     "city": "ABC 3 Lafayette LA KATC 720p",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 30.2241,
-    "geo_long": -92.0198,
+    "geo_lat": 30.2262187,
+    "geo_long": -92.0178202,
     "latitude": 30.2241,
     "longitude": -92.0198
   },
@@ -666,8 +666,8 @@
     "city": "ABC 25 News Central Texas",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 37.0902,
-    "geo_long": -95.7129,
+    "geo_lat": 39.7837304,
+    "geo_long": -100.445882,
     "latitude": 37.0902,
     "longitude": -95.7129
   },
@@ -1044,8 +1044,8 @@
     "city": "Adom",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 0,
-    "geo_long": 0,
+    "geo_lat": 7.4687328,
+    "geo_long": -2.5983739,
     "latitude": 0,
     "longitude": 0
   },
@@ -3858,8 +3858,8 @@
     "city": "Bolton Community Voice Channel",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 50,
-    "geo_long": -46,
+    "geo_lat": 41.7689878,
+    "geo_long": -72.4334173,
     "latitude": 50,
     "longitude": -46
   },
@@ -4222,8 +4222,8 @@
     "city": "Canal 2 International",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 0,
-    "geo_long": 0,
+    "geo_lat": 4.6125522,
+    "geo_long": 13.1535811,
     "latitude": 0,
     "longitude": 0
   },
@@ -5034,8 +5034,8 @@
     "city": "CAPS Media Channel 6 Ventura C",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 58,
-    "geo_long": -30,
+    "geo_lat": 34.2783355,
+    "geo_long": -119.293174,
     "latitude": 58,
     "longitude": -30
   },
@@ -5048,8 +5048,8 @@
     "city": "CAPS Media Channel 15 Ventura ",
     "categories": "General",
     "type": "tv",
-    "geo_lat": 40,
-    "geo_long": -28,
+    "geo_lat": 34.2783355,
+    "geo_long": -119.293174,
     "latitude": 40,
     "longitude": -28
   },
